Add real time clock timestamping for portable mode. (#52)
The way this works is the PC updates the BattOr's clock every time the battor sw is run. Since the BattOr is always powered on when connected to a test device, the internal ms timer just keeps the RTC time up to date.  When a portable trace is run the start of the trace is stored with the trace file in the fs.

+++ fw/drivers/button.c
@@ -7,7 +7,7 @@
 #include "button.h"
 
 static uint16_t down_ms;
-static uint16_t prev_ms_ticks;
+static uint32_t prev_ms_ticks;
 static uint8_t button_state;
 typedef enum BUTTON_STATE_enum
 {
